third_party/blink/web_tests/http/tests/devtools/sources/debugger-console/debug-console-command-expected.txt

Tests debug(fn) console command.


Running: testSetSimpleBreakpoint
Breakpoint added.
Set timer for test function.
Script execution paused.
Call stack:
    0) simpleTestFunction (debug-console-command.js:15)
    1)  (:1)
Breakpoint removed.
Script execution resumed.

Running: testSetBreakpointOnFirstFunctionInLine
Breakpoint added.
Set timer for test function.
Script execution paused.
Call stack:
    0) simpleTestFunction1 (debug-console-command.js:19)
    1)  (:1)
Breakpoint removed.
Script execution resumed.

Running: testSetBreakpointOnLastFunctionInLine
Breakpoint added.
Set timer for test function.
Script execution paused.
Call stack:
    0) simpleTestFunction2 (debug-console-command.js:19)
    1)  (:1)
Breakpoint removed.
Script execution resumed.

Running: testRemoveBreakpoint
Script execution paused.
Reason for pause: debugger statement.
